Refactor and renamed attributes in TypicalPersons.jsva
