+++ src/main/java/seedu/address/logic/commands/ClearCommand.java
@@ -12,7 +12,7 @@ import seedu.address.model.TaskCollection;
 public class ClearCommand extends Command {
 
     public static final String COMMAND_WORD = "clear";
-    public static final String MESSAGE_SUCCESS = "deadline manager has been cleared!";
+    public static final String MESSAGE_SUCCESS = "Deadline manager has been cleared!";
 
 
     @Override
