+++ src/test/java/seedu/address/logic/commands/AttachmentCommandTest.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_TASK;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalTasks.getTypicalTaskCollections;
 
 import java.io.File;
 import java.io.IOException;
@@ -30,7 +30,7 @@ import seedu.address.testutil.PersonBuilder;
 public class AttachmentCommandTest {
     private static File nonExistentFile = new File("blabla/not-here1827364.txt");
 
-    private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+    private Model model = new ModelManager(getTypicalTaskCollections(), new UserPrefs());
     private CommandHistory commandHistory = new CommandHistory();
 
     /**
