+++ src/test/java/seedu/address/model/TaskCollectionTest.java
@@ -5,8 +5,8 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_PRIORITY_BOB;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalTasks.ALICE;
+import static seedu.address.testutil.TypicalTasks.getTypicalTaskCollections;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -42,7 +42,7 @@ public class TaskCollectionTest {
 
     @Test
     public void resetData_withValidReadOnlyAddressBook_replacesData() {
-        TaskCollection newData = getTypicalAddressBook();
+        TaskCollection newData = getTypicalTaskCollections();
         taskCollection.resetData(newData);
         assertEquals(newData, taskCollection);
     }
